+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Allow specifying which axes are included in the final measurement when using `ElementContent(measuring:)`.
+
 ### Removed
 
 ### Changed